+++ docker/scripts/make_directories.sh
@@ -1,0 +1,7 @@
+#!/usr/bin/env bash
+
+# Make directories we need before running the rest of the scripts
+mkdir -p \
+  /deploy/server/www \
+  /deploy/server/logs/apache \
+  /deploy/server/logs/launcher