+++ examples/PlainPython/Plotly/basic/app.py
@@ -55,9 +55,13 @@ PLOTS = {
     "Scatter": scatter,
 }
 
+def on_event(type, e):
+    print(type, e)
+
 
 html_plot = None
-layout = SinglePage("Plotly demo")
+layout = SinglePage("Plotly")
+layout.title.set_text("trame ❤️ plotly")
 
 with layout.toolbar:
     vuetify.VSpacer()
@@ -69,8 +73,18 @@ with layout.toolbar:
     )
 
 with layout.content:
-    with vuetify.VContainer(classes="fluid fill-height"):
-        html_plot = plotly.Plotly("demo", display_mode_bar=True)
+    with vuetify.VContainer(fluid=True):
+        with vuetify.VRow(dense=True):
+            vuetify.VSpacer()
+            html_plot = plotly.Plotly(
+                "demo",
+                display_mode_bar=("true",),
+                selected=(on_event, "['selected', VuePlotly.safe($event)]"),
+                # hover=(on_event, "['hover', VuePlotly.safe($event)]"),
+                # selecting=(on_event, "['selecting', $event]"),
+                # unhover=(on_event, "['unhover', $event]"),
+            )
+            vuetify.VSpacer()
 
 
 @state.change("active_plot")