+++ trame/__init__.py
@@ -1,6 +1,7 @@
 from trame.internal import (
   change, Controller, flush_state, get_cli_parser, get_state, get_version,
-  is_dirty, is_dirty_all, port, start, State, stop, trigger, update_state
+  is_dirty, is_dirty_all, port, setup_dev, start, State, stop, trigger,
+  update_state, RemoteFile
 )
 from trame.layouts import update_layout
 
@@ -79,6 +80,12 @@ __all__ = [
 
     # CLI-related
     "get_cli_parser",
+
+    # Dev-related
+    "setup_dev",
+
+    # Utils/helpers
+    "RemoteFile",
 
     # These are not exposed in the docs
     "__version__",