+++ trame/html/plotly/__init__.py
@@ -39,7 +39,7 @@ class Plotly(AbstractElement):
     Create a Plotly figure element
     """
 
-    def __init__(self, _name, **kwargs):
+    def __init__(self, _name, figure=None, **kwargs):
         super().__init__(
             "Plotly",
             data=(f"{_name}_figure.data",),
@@ -47,6 +47,7 @@ class Plotly(AbstractElement):
             **kwargs,
         )
         self.__figure_key = f"{_name}_figure"
+        self.__figure_data = figure
         state[self.__figure_key] = {"data": [], "layout": {}}
         self._attr_names += [
             "data",
@@ -67,6 +68,43 @@ class Plotly(AbstractElement):
             "responsive",
             ("double_click_delay", "doubleClickDelay"),
         ]
+        self._event_names += [
+            ("after_export", "afterexport"),
+            ("after_plot", "afterplot"),
+            ("animated", "animated"),
+            ("animating_frame", "animatingframe"),
+            ("animation_interrupted", "animationinterrupted"),
+            ("auto_size", "autosize"),
+            ("before_export", "beforeexport"),
+            ("button_clicked", "buttonclicked"),
+            ("click", "click"),
+            ("click_annotation", "clickannotation"),
+            ("deselect", "deselect"),
+            ("double_click", "doubleclick"),
+            ("framework", "framework"),
+            ("hover", "hover"),
+            ("legend_click", "legendclick"),
+            ("legend_double_click", "legenddoubleclick"),
+            ("relayout", "relayout"),
+            ("restyle", "restyle"),
+            ("redraw", "redraw"),
+            ("selected", "selected"),
+            ("selecting", "selecting"),
+            ("slider_change", "sliderchange"),
+            ("slider_end", "sliderend"),
+            ("slider_start", "sliderstart"),
+            ("transitioning", "transitioning"),
+            ("transition_interrupted", "transitioninterrupted"),
+            ("unhover", "unhover"),
+        ]
+        self.update()
 
-    def update(self, plotly_fig):
-        state[self.__figure_key] = safe_figure(plotly_fig.to_plotly_json())
+    def update(self, plotly_fig=None):
+        if plotly_fig:
+            self.__figure_data = plotly_fig
+
+        if self.__figure_data:
+            state[self.__figure_key] = safe_figure(self.__figure_data.to_plotly_json())
+
+class Figure(Plotly):
+    pass